fix: go1.19 specific tests for text wrapping
